requirements.txt
django==5.2.7
djangorestframework>=3.14
django-cors-headers>=4.3.1
gunicorn>=20.1.0
liac-arff